create testing module and move random.py in there
