+++ test/circuit/library/test_bogoliubov_transform.py
@@ -13,7 +13,6 @@
 """Test Bogoliubov transform circuits."""
 
 from test import QiskitNatureDeprecatedTestCase
-from test.nature_random import random_legacy_quadratic_hamiltonian as random_quadratic_hamiltonian
 
 import numpy as np
 from ddt import data, ddt, unpack
@@ -24,6 +23,9 @@ from qiskit_nature.circuit.library import BogoliubovTransform
 from qiskit_nature.converters.second_quantization import QubitConverter
 from qiskit_nature.mappers.second_quantization import BravyiKitaevMapper, JordanWignerMapper
 from qiskit_nature.operators.second_quantization.quadratic_hamiltonian import QuadraticHamiltonian
+from qiskit_nature.testing.random import (
+    random_legacy_quadratic_hamiltonian as random_quadratic_hamiltonian,
+)
 
 
 def _expand_transformation_matrix(mat: np.ndarray) -> np.ndarray:
