+++ test/operators/second_quantization/test_quadratic_hamiltonian.py
@@ -13,7 +13,6 @@
 """Tests for QuadraticHamiltonian"""
 
 from test import QiskitNatureDeprecatedTestCase
-from test.nature_random import random_antisymmetric_matrix
 
 import numpy as np
 from ddt import data, ddt
@@ -23,6 +22,7 @@ from qiskit_nature.converters.second_quantization import QubitConverter
 from qiskit_nature.mappers.second_quantization import JordanWignerMapper
 from qiskit_nature.operators.second_quantization import QuadraticHamiltonian
 from qiskit_nature.operators.second_quantization.fermionic_op import FermionicOp
+from qiskit_nature.testing import random_antisymmetric_matrix
 
 
 @ddt
