+++ test/second_q/circuit/library/test_bogoliubov_transform.py
@@ -13,7 +13,6 @@
 """Test Bogoliubov transform circuits."""
 
 from test import QiskitNatureTestCase
-from test.nature_random import random_quadratic_hamiltonian
 
 import numpy as np
 from ddt import data, ddt, unpack
@@ -21,14 +20,9 @@ from qiskit import QuantumCircuit, QuantumRegister
 from qiskit.quantum_info import Operator, Statevector, random_hermitian, random_unitary
 
 from qiskit_nature.second_q.circuit.library import BogoliubovTransform
-from qiskit_nature.second_q.mappers import QubitConverter
-from qiskit_nature.second_q.mappers import (
-    BravyiKitaevMapper,
-    JordanWignerMapper,
-)
-from qiskit_nature.second_q.hamiltonians.quadratic_hamiltonian import (
-    QuadraticHamiltonian,
-)
+from qiskit_nature.second_q.hamiltonians.quadratic_hamiltonian import QuadraticHamiltonian
+from qiskit_nature.second_q.mappers import BravyiKitaevMapper, JordanWignerMapper, QubitConverter
+from qiskit_nature.testing import random_quadratic_hamiltonian
 
 
 def _expand_transformation_matrix(mat: np.ndarray) -> np.ndarray:
